+++ tests/Functional/Resources/Reports/test_ReportsClient.py
@@ -8,6 +8,7 @@ from paddle_billing.Entities.ReportCSV import ReportCSV
 from paddle_billing.Entities.Reports import (
     AdjustmentsReportType,
     BalanceReportType,
+    CheckoutsReportType,
     DiscountsReportType,
     PayoutReconciliationReportType,
     ProductsPricesReportType,
@@ -23,8 +24,10 @@ from paddle_billing.Entities.DateTime import DateTime
 from paddle_billing.Resources.Reports.Operations.Filters import (
     AdjustmentActionFilter,
     AdjustmentStatusFilter,
+    CheckoutCreatedAtFilter,
     CollectionModeFilter,
     CurrencyCodeFilter,
+    CustomerCountryCodeFilter,
     DiscountStatusFilter,
     DiscountTypeFilter,
     PriceStatusFilter,
@@ -45,6 +48,7 @@ from paddle_billing.Entities.Shared import (
     Action,
     CatalogType,
     CollectionMode,
+    CountryCode,
     CurrencyCode,
     TransactionOrigin,
     TransactionStatus,
@@ -61,6 +65,7 @@ from paddle_billing.Entities.Discounts import (
 from paddle_billing.Resources.Reports.Operations import (
     CreateAdjustmentsReport,
     CreateBalanceReport,
+    CreateCheckoutsReport,
     CreateDiscountsReport,
     CreatePayoutReconciliationReport,
     CreateProductsAndPricesReport,
@@ -174,6 +179,23 @@ class TestReportsClient:
                 ReadsFixtures.read_raw_json_fixture("request/create_payout_reconciliation_full"),
                 PayoutReconciliationReportType,
             ),
+            (
+                CreateCheckoutsReport(type=CheckoutsReportType.Checkouts),
+                ReadsFixtures.read_raw_json_fixture("request/create_checkouts_basic"),
+                CheckoutsReportType,
+            ),
+            (
+                CreateCheckoutsReport(
+                    type=CheckoutsReportType.Checkouts,
+                    filters=[
+                        CheckoutCreatedAtFilter.gte(DateTime("2023-12-30")),
+                        CheckoutCreatedAtFilter.lt(DateTime("2024-12-30T12:30:01.123456Z")),
+                        CustomerCountryCodeFilter([CountryCode.US, CountryCode.GB]),
+                    ],
+                ),
+                ReadsFixtures.read_raw_json_fixture("request/create_checkouts_full"),
+                CheckoutsReportType,
+            ),
         ],
         ids=[
             "Create transaction report with basic data",
@@ -185,6 +207,8 @@ class TestReportsClient:
             "Create balance report with filters",
             "Create payout reconciliation report with basic data",
             "Create payout reconciliation report with filters",
+            "Create checkouts report with basic data",
+            "Create checkouts report with filters",
         ],
     )
     def test_create_report_uses_expected_payload(